endpoint: Defer 1st labels update until 1st policy realization
Start a controller to delay the update of the labels until we have at
least generated completely one policy for the endpoint.
This allows configuring the init policy for the endpoint completely, in
order to allow enough connectivity for a sidecar proxy to get
configured, which can be required to enforce L7 policy rules as a result
of the labels update.

Start the repository revision from 1, so that when an Endpoint policy
is generated the endpoint has a policyRevision > 0, which allows
identifying when an endpoint has had any policy generated.

Signed-off-by: Romain Lenglet <romain@covalent.io>